requirements.txt
PyJWT==2.8.0
requests==2.31.0
aiohttp==3.8.6
PyJWT==1.7.1
Django==3.1.5
django-cors-headers==3.8.0
djangorestframework==3.12.2
djangorestframework-simplejwt==5.2.2
